End of user crud functinality

+++ Theater/src/main/java/com/epam/theater/dao/UserDao.java
@@ -3,5 +3,5 @@ package com.epam.theater.dao;
 import com.epam.theater.bean.User;
 
 public interface UserDao extends AbstractDomainObjectDao<User> {
-	int getFreeId();
+	Long getFreeId();
 }
